Refactored DuplicationResults into DuplicateInstance and VisualizeDiffs to scope down to comparing pairs of code clones, ignore diffs in whitespace and reordered lines in the diff highlighting
