+++ CodeDuplicationChecker/CodeDuplicationChecker.csproj
@@ -47,8 +47,8 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CodeIterator.cs" />
-    <Compile Include="DuplicationResults.cs" />
-    <Compile Include="MyClass.cs" />
+    <Compile Include="DuplicateInstance.cs" />
+    <Compile Include="VisualizeDiffs.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SimilarityScorer.cs" />
