--- CodeDuplicationChecker/MyClass.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace CodeDuplicationChecker
-{
-    public class MyClass
-    {
-        public string HelloWorld()
-        {
-            return "Hello world";
-        }
-    }
-}
