+++ CodeDuplicationCheckerTests/CodeDuplicationCheckerTests.csproj
@@ -50,8 +50,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CodeIteratorTests.cs" />
-    <Compile Include="DuplicationResultsTests.cs" />
-    <Compile Include="MyClassTests.cs" />
+    <Compile Include="VisualizeDiffsTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SampleCode\RiddledWithDuplicates.cs" />
     <Compile Include="SimilarityScorerTests.cs" />
