+++ CodeDuplicationCheckerTests/VisualizeDiffsTests.cs
@@ -6,43 +6,54 @@ namespace CodeDuplicationChecker.Tests
     [TestClass()]
     public class DuplicationResultsTests
     {
-        DuplicationResults duplicationResults;
-
-        [TestInitialize]
-        public void TestInitialize()
+        [TestMethod()]
+        public void GenerateResultsFileTest1()
         {
-            duplicationResults = new DuplicationResults();
+            // Act
+            VisualizeDiffs.GenerateResultsFile(type1, false);
+
+            // Assert
+
         }
 
         [TestMethod()]
-        public void GenerateResultsFileTest()
+        public void GenerateResultsFileTest2()
         {
-            // Arrange
-            duplicationResults.CodeDuplicates = new List<DuplicateCode>()
-            {
-                type1.Copy(),
-                type2.Copy(),
-                type3.Copy()
-            };
-
             // Act
-            duplicationResults.GenerateResultsFile(false);
+            VisualizeDiffs.GenerateResultsFile(type2, false);
+
+            // Assert
+
+        }
+
+        [TestMethod()]
+        public void GenerateResultsFileTest3()
+        {
+            // Act
+            VisualizeDiffs.GenerateResultsFile(type3, false);
+
+            // Assert
+
         }
 
         [TestMethod()]
         public void GenerateCodeHTML_Type1_Test()
         {
             // Arrange
-            var copy = type1.Copy();
+            var copy = new List<DuplicateInstance>();
+            foreach (var instance in type1)
+            {
+                copy.Add(instance.Copy());
+            }
 
             // Act
-            copy.GenerateCodeHTML();
+            VisualizeDiffs.GenerateCodeHTML(copy);
 
             // Assert
-            for (int i = 0; i < type1.Instances.Count; i++)
+            for (int i = 0; i < type1.Count; i++)
             {
-                var expected = type2.Instances[i].CodeHtml.Trim();
-                var actual = copy.Instances[i].CodeHtml.Trim();
+                var expected = type1[i].CodeHtml.Trim();
+                var actual = copy[i].CodeHtml.Trim();
                 Assert.AreEqual(expected, actual, $"Instance {i} failed");
             }
         }
@@ -51,16 +62,20 @@ namespace CodeDuplicationChecker.Tests
         public void GenerateCodeHTML_Type2_Test()
         {
             // Arrange
-            var copy = type2.Copy();
+            var copy = new List<DuplicateInstance>();
+            foreach (var instance in type2)
+            {
+                copy.Add(instance.Copy());
+            }
 
             // Act
-            copy.GenerateCodeHTML();
+            VisualizeDiffs.GenerateCodeHTML(copy);
 
             // Assert
-            for (int i = 0; i < type2.Instances.Count; i++)
+            for (int i = 0; i < type2.Count; i++)
             {
-                var expected = type2.Instances[i].CodeHtml.Trim();
-                var actual = copy.Instances[i].CodeHtml.Trim();
+                var expected = type2[i].CodeHtml.Trim();
+                var actual = copy[i].CodeHtml.Trim();
                 Assert.AreEqual(expected, actual, $"Instance {i} failed");
             }
         }
@@ -69,125 +84,34 @@ namespace CodeDuplicationChecker.Tests
         public void GenerateCodeHTML_Type3_Test()
         {
             // Arrange
-            var copy = type3.Copy();
+            // Arrange
+            var copy = new List<DuplicateInstance>();
+            foreach (var instance in type3)
+            {
+                copy.Add(instance.Copy());
+            }
 
             // Act
-            copy.GenerateCodeHTML();
+            VisualizeDiffs.GenerateCodeHTML(copy);
 
             // Assert
-            for (int i = 0; i < type3.Instances.Count; i++)
+            for (int i = 0; i < type3.Count; i++)
             {
-                var expected = type2.Instances[i].CodeHtml.Trim();
-                var actual = copy.Instances[i].CodeHtml.Trim();
+                var expected = type3[i].CodeHtml.Trim();
+                var actual = copy[i].CodeHtml.Trim();
                 Assert.AreEqual(expected, actual, $"Instance {i} failed");
             }
         }
 
-        [TestMethod()]
-        public void Master_Type1_Test()
-        {
-            // Arrange
-            var copy = type1.Copy();
-            var expected = @"if (!loadedInstances.ContainsKey(action.TestClass))
-{
-    // Get the Constructor
-    ConstructorInfo methodConstructor = type.GetConstructor(Type.EmptyTypes);
-    methodInstance = methodConstructor.Invoke(new object[] { });
-    // Need to call SetTestContext(TestContext context) to pass test item setting and logger info. And use same test context for all the action modules in same test.
-    methodInfo = type.GetMethod(""SetTestContext"");
-    object[] setTextContextParameters = new object[1];
-    setTextContextParameters[0] = testContext;
-    methodInfo.Invoke(methodInstance, setTextContextParameters);
-    loadedInstances.Add(action.TestClass, methodInstance);
-}
-else
-{
-    methodInstance = loadedInstances[action.TestClass];
-}";
-
-            // Act
-            var actual = copy.Master;
-
-            // Assert
-            Assert.AreEqual(expected, actual);
-        }
-
-        [TestMethod()]
-        public void Master_Type2_Test()
-        {
-            // Arrange
-            var copy = type2.Copy();
-            var expected = @"/// <summary>
-/// Update the RunTime Value in the *Tests.xml 
-/// </summary>
-/// <param name=""templateDefintion""></param>
-/// <returns></returns>
-public static List<TestItem.Action> UpdateActionsRunTimeValues(List<TestItem.Action> templateDefintion, TestContext testContext)
-{
-    XmlSerializer s = new XmlSerializer(typeof(List<TestItem.Action>));
-    using (StringWriter outStream = new StringWriter())
-    {
-        s.Serialize(outStream, templateDefintion);
-        string input = outStream.ToString();
-        input = ReplaceRunTimeValues(input, testContext, ref s);
-        using (StringReader rdr = new StringReader(input))
-        {
-            templateDefintion = (List<TestItem.Action>)s.Deserialize(rdr);
-        }
-    }
-    return templateDefintion;
-}";
-
-            // Act
-            var actual = copy.Master;
-
-            // Assert
-            Assert.AreEqual(expected, actual);
-        }
-
-        [TestMethod()]
-        public void Master_Type3_Test()
-        {
-            // Arrange
-            var copy = type3.Copy();
-            var expected = @"public static List<TestItem.Action> UpdateActionsRunTimeValues(List<TestItem.Action> templateDefintion, TestContext testContext)
-{
-    XmlSerializer s = new XmlSerializer(typeof(List<TestItem.Action>));
-    using (StringWriter outStream = new StringWriter())
-    {
-        // Here are some reordered lines
-        int j = 5 * 6;
-        int i = 3 * 5;
-        // This is an added comment
-        s.Serialize(outStream, templateDefintion);
-        string input = outStream.ToString();
-        input = ReplaceRunTimeValues(input, testContext, ref s);
-        using (StringReader rdr = new StringReader(input))
-        {
-            templateDefintion = (List<TestItem.Action>)s.Deserialize(rdr);
-        }
-    }
-    return templateDefintion;
-}";
-
-            // Act
-            var actual = copy.Master;
-
-            // Assert
-            Assert.AreEqual(expected, actual);
-        }
-
         #region TestCode
-        private readonly DuplicateCode type1 = new DuplicateCode()
+        private readonly List<DuplicateInstance> type1 = new List<DuplicateInstance>()
         {
-            Instances = new List<DuplicateInstance>()
+            new DuplicateInstance ()
             {
-                new DuplicateInstance ()
-                {
-                    Filename = "file1.cs",
-                    StartLine = 80,
-                    EndLine = 97,
-                    Code = @"
+                Filename = "file1.cs",
+                StartLine = 80,
+                EndLine = 97,
+                Code = @"
 if (!loadedInstances.ContainsKey(action.TestClass))
 {
     // Get the Constructor
@@ -206,7 +130,7 @@ else
 {
     methodInstance = loadedInstances[action.TestClass];
 }",
-                    CodeHtml =@"
+                CodeHtml =@"
 <span class='same'>if (!loadedInstances.ContainsKey(action.TestClass))</span>
 <span class='same'>{</span>
 <span class='same'>    // Get the Constructor</span>
@@ -225,19 +149,18 @@ else
 <span class='same'>{</span>
 <span class='same'>    methodInstance = loadedInstances[action.TestClass];</span>
 <span class='same'>}</span>",
-                },
-                new DuplicateInstance ()
-                {
-                    Filename = "file1.cs",
-                    StartLine = 100,
-                    EndLine = 117,
-                    Code = @"
+            },
+            new DuplicateInstance ()
+            {
+                Filename = "file1.cs",
+                StartLine = 100,
+                EndLine = 117,
+                Code = @"
 if (!loadedInstances.ContainsKey(action.TestClass))
 {
     // Get the Constructor
     ConstructorInfo methodConstructor = type.GetConstructor(Type.EmptyTypes);
     methodInstance = methodConstructor.Invoke(new object[] { });
-
 
     // Need to call SetTestContext(TestContext context) to pass test item setting and logger info. And use same test context for all the action modules in same test.
     methodInfo = type.GetMethod(""SetTestContext"");
@@ -250,14 +173,14 @@ else
 {
     methodInstance = loadedInstances[action.TestClass];
 }",
-                    CodeHtml = @"
+                CodeHtml = @"
 <span class='same'>if (!loadedInstances.ContainsKey(action.TestClass))</span>
 <span class='same'>{</span>
 <span class='same'>    // Get the Constructor</span>
 <span class='same'>    ConstructorInfo methodConstructor = type.GetConstructor(Type.EmptyTypes);</span>
 <span class='same'>    methodInstance = methodConstructor.Invoke(new object[] { });</span>
 <span class='same'></span>
-<span class='diff'></span>
+<span class='same'></span>
 <span class='same'>    // Need to call SetTestContext(TestContext context) to pass test item setting and logger info. And use same test context for all the action modules in same test.</span>
 <span class='same'>    methodInfo = type.GetMethod(""SetTestContext"");</span>
 <span class='same'>    object[] setTextContextParameters = new object[1];</span>
@@ -270,63 +193,17 @@ else
 <span class='same'>{</span>
 <span class='same'>    methodInstance = loadedInstances[action.TestClass];</span>
 <span class='same'>}</span>",
-                },
-                new DuplicateInstance ()
-                {
-                    Filename = "file2.cs",
-                    StartLine = 55,
-                    EndLine = 72,
-                    Code = @"
-if (!loadedInstances.ContainsKey(action.TestClass))
-{
-    // Get the Constructor
-    ConstructorInfo methodConstructor = type.GetConstructor(Type.EmptyTypes);
-    methodInstance = methodConstructor.Invoke(new object[] { });
-
-    // Need to call SetTestContext(TestContext context) to pass test item setting and logger info. And use same test context for all the action modules in same test.
-    methodInfo = type.GetMethod(""SetTestContext"");
-    object[] setTextContextParameters = new object[1];
-    setTextContextParameters[0] = testContext;
-    methodInfo.Invoke(methodInstance, setTextContextParameters);
-    loadedInstances.Add(action.TestClass, methodInstance);
-}
-else
-{
-    methodInstance = loadedInstances[action.TestClass];
-}",
-                    CodeHtml = @"
-<span class='same'>if (!loadedInstances.ContainsKey(action.TestClass))</span>
-<span class='same'>{</span>
-<span class='same'>    // Get the Constructor</span>
-<span class='same'>    ConstructorInfo methodConstructor = type.GetConstructor(Type.EmptyTypes);</span>
-<span class='same'>    methodInstance = methodConstructor.Invoke(new object[] { });</span>
-<span class='same'></span>
-<span class='same'>    // Need to call SetTestContext(TestContext context) to pass test item setting and logger info. And use same test context for all the action modules in same test.</span>
-<span class='same'>    methodInfo = type.GetMethod(""SetTestContext"");</span>
-<span class='same'>    object[] setTextContextParameters = new object[1];</span>
-<span class='same'>    setTextContextParameters[0] = testContext;</span>
-<span class='diff'></span>
-<span class='same'>    methodInfo.Invoke(methodInstance, setTextContextParameters);</span>
-<span class='same'>    loadedInstances.Add(action.TestClass, methodInstance);</span>
-<span class='same'>}</span>
-<span class='same'>else</span>
-<span class='same'>{</span>
-<span class='same'>    methodInstance = loadedInstances[action.TestClass];</span>
-<span class='same'>}</span>",
-                }
             }
         };
 
-        private readonly DuplicateCode type2 = new DuplicateCode()
+        private readonly List<DuplicateInstance> type2 = new List<DuplicateInstance>()
         {
-            Instances = new List<DuplicateInstance>()
+            new DuplicateInstance ()
             {
-                new DuplicateInstance ()
-                {
-                    Filename = "file5.cs",
-                    StartLine = 14,
-                    EndLine = 35,
-                    Code = @"
+                Filename = "file5.cs",
+                StartLine = 14,
+                EndLine = 35,
+                Code = @"
 /// <summary>
 /// Update the RunTime Value in the *Tests.xml 
 /// </summary>
@@ -348,7 +225,7 @@ public static List<TestItem.Action> UpdateActionsRunTimeValues(List<TestItem.Act
     }
     return templateDefintion;
 }",
-                    CodeHtml = @"
+                CodeHtml = @"
 <span class='same'>/// &lt;summary&gt;</span>
 <span class='same'>/// Update the RunTime Value in the *Tests.xml</span>
 <span class='same'>/// &lt;/summary&gt;</span>
@@ -370,13 +247,13 @@ public static List<TestItem.Action> UpdateActionsRunTimeValues(List<TestItem.Act
 <span class='same'>    }</span>
 <span class='same'>    return templateDefintion;</span>
 <span class='same'>}</span>",
-                },
-                new DuplicateInstance ()
-                {
-                    Filename = "file5.cs",
-                    StartLine = 40,
-                    EndLine = 61,
-                    Code = @"
+            },
+            new DuplicateInstance ()
+            {
+                Filename = "file5.cs",
+                StartLine = 40,
+                EndLine = 61,
+                Code = @"
 /// <summary>
 /// Update the RunTime Value in the *Tests.xml 
 /// </summary>
@@ -398,7 +275,7 @@ public static TestItem.TemplateItemsDefinition UpdateTemplatesRunTimeValues(Test
     }
     return templateDefintion;
 }",
-                    CodeHtml = @"
+                CodeHtml = @"
 <span class='same'>/// &lt;summary&gt;</span>
 <span class='same'>/// Update the RunTime Value in the *Tests.xml</span>
 <span class='same'>/// &lt;/summary&gt;</span>
@@ -420,41 +297,38 @@ public static TestItem.TemplateItemsDefinition UpdateTemplatesRunTimeValues(Test
 <span class='same'>    }</span>
 <span class='same'>    return templateDefintion;</span>
 <span class='same'>}</span>",
-                }
             }
-};
+        };
 
-        private readonly DuplicateCode type3 = new DuplicateCode()
+        private readonly List<DuplicateInstance> type3 = new List<DuplicateInstance>()
         {
-            Instances = new List<DuplicateInstance>()
+            new DuplicateInstance ()
             {
-                new DuplicateInstance ()
-                {
-                    Filename = "file9.cs",
-                    StartLine = 14,
-                    EndLine = 35,
-                    Code = @"
+                Filename = "file9.cs",
+                StartLine = 14,
+                EndLine = 35,
+                Code = @"
 public static List<TestItem.Action> UpdateActionsRunTimeValues(List<TestItem.Action> templateDefintion, TestContext testContext)
 {
-    XmlSerializer s = new XmlSerializer(typeof(List<TestItem.Action>));
+XmlSerializer s = new XmlSerializer(typeof(List<TestItem.Action>));
 
-    using (StringWriter outStream = new StringWriter())
+using (StringWriter outStream = new StringWriter())
+{
+    // Here are some reordered lines
+    int j = 5 * 6;
+    int i = 3 * 5;
+    // This is an added comment
+    s.Serialize(outStream, templateDefintion);
+    string input = outStream.ToString();
+    input = ReplaceRunTimeValues(input, testContext, ref s);
+    using (StringReader rdr = new StringReader(input))
     {
-        // Here are some reordered lines
-        int j = 5 * 6;
-        int i = 3 * 5;
-        // This is an added comment
-        s.Serialize(outStream, templateDefintion);
-        string input = outStream.ToString();
-        input = ReplaceRunTimeValues(input, testContext, ref s);
-        using (StringReader rdr = new StringReader(input))
-        {
-            templateDefintion = (List<TestItem.Action>)s.Deserialize(rdr);
-        }
+        templateDefintion = (List<TestItem.Action>)s.Deserialize(rdr);
     }
-    return templateDefintion;
+}
+return templateDefintion;
 }",
-                    CodeHtml = @"
+                CodeHtml = @"
 <span class='same'>public static List<TestItem.Action> UpdateActionsRunTimeValues(List<TestItem.Action> templateDefintion, TestContext testContext)</span>
 <span class='same'>{</span>
 <span class='same'>    XmlSerializer s = new XmlSerializer(typeof(List<TestItem.Action>));</span>
@@ -462,8 +336,8 @@ public static List<TestItem.Action> UpdateActionsRunTimeValues(List<TestItem.Act
 <span class='same'>    using (StringWriter outStream = new StringWriter())</span>
 <span class='same'>    {</span>
 <span class='same'>        // Here are some reordered lines</span>
-<span class='diff'>        int j = 5 * 6;</span>
-<span class='diff'>        int i = 3 * 5;</span>
+<span class='same'>        int j = 5 * 6;</span>
+<span class='same'>        int i = 3 * 5;</span>
 <span class='diff'>        // This is an added line</span>
 <span class='same'>        s.Serialize(outStream, templateDefintion);</span>
 <span class='same'>        string input = outStream.ToString();</span>
@@ -475,34 +349,34 @@ public static List<TestItem.Action> UpdateActionsRunTimeValues(List<TestItem.Act
 <span class='same'>    }</span>
 <span class='same'>    return templateDefintion;</span>
 <span class='same'>}</span>"
-                },
-                new DuplicateInstance ()
-                {
-                    Filename = "file7.cs",
-                    StartLine = 40,
-                    EndLine = 61,
-                    Code = @"
+            },
+            new DuplicateInstance ()
+            {
+                Filename = "file7.cs",
+                StartLine = 40,
+                EndLine = 61,
+                Code = @"
 public static TestItem.TemplateItemsDefinition UpdateTemplatesRunTimeValues(TestItem.TemplateItemsDefinition templateDefintion, TestContext testContext)
 {
-    XmlSerializer s = new XmlSerializer(typeof(TestItem.TemplateItemsDefinition));
+XmlSerializer s = new XmlSerializer(typeof(TestItem.TemplateItemsDefinition));
 
-    using (StringWriter outStream = new StringWriter())
+using (StringWriter outStream = new StringWriter())
+{
+    // Here are some reordered lines
+    int i = 3 * 5;
+    int j = 5 * 6;
+    s.Serialize(outStream, templateDefintion);
+    string input = outStream.ToString();
+    // Here I have removed a line
+    using (StringReader rdr = new StringReader(input))
     {
-        // Here are some reordered lines
-        int i = 3 * 5;
-        int j = 5 * 6;
-        s.Serialize(outStream, templateDefintion);
-        string input = outStream.ToString();
-        // Here I have removed a line
-        using (StringReader rdr = new StringReader(input))
-        {
-            templateDefintion = (TestItem.TemplateItemsDefinition)s.Deserialize(rdr);
-        }
+        templateDefintion = (TestItem.TemplateItemsDefinition)s.Deserialize(rdr);
     }
+}
 
-    return templateDefintion;
+return templateDefintion;
 }",
-                    CodeHtml = @"
+                CodeHtml = @"
 <span class='same'>public static List<TestItem.Action> UpdateActionsRunTimeValues(List<TestItem.Action> templateDefintion, TestContext testContext)</span>
 <span class='same'>{</span>
 <span class='same'>    XmlSerializer s = new XmlSerializer(typeof(List<TestItem.Action>));</span>
@@ -510,8 +384,8 @@ public static TestItem.TemplateItemsDefinition UpdateTemplatesRunTimeValues(Test
 <span class='same'>    using (StringWriter outStream = new StringWriter())</span>
 <span class='same'>    {</span>
 <span class='same'>        // Here are some reordered lines</span>
-<span class='diff'>        int i = 3 * 5;</span>
-<span class='diff'>        int j = 5 * 6;</span>
+<span class='same'>        int i = 3 * 5;</span>
+<span class='same'>        int j = 5 * 6;</span>
 <span class='same'>        s.Serialize(outStream, templateDefintion);</span>
 <span class='same'>        string input = outStream.ToString();</span>
 <span class='diff'>        // Here I have removed a line</span>
@@ -522,7 +396,6 @@ public static TestItem.TemplateItemsDefinition UpdateTemplatesRunTimeValues(Test
 <span class='same'>    }</span>
 <span class='same'>    return templateDefintion;</span>
 <span class='same'>}</span>"
-                }
             }
         };
         #endregion
